fix: extract pay data types to contracts/ tier — fix storage→bricks violation
Move pure data types (SpendingPolicy, SpendingApproval, SpendingLedgerEntry,
PolicyEvaluation) and conversion functions (credits_to_micro, micro_to_credits,
MICRO_UNIT_SCALE) from bricks/pay/ to contracts/pay_types.py.

storage/repositories/spending_policy.py now imports directly from contracts/
instead of using importlib to bypass the tier boundary (storage cannot import
from bricks). bricks/pay/ re-exports all moved symbols for backward compat.
